fix endian
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@825 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/engine/midi_mtrk.cpp
@@ -326,7 +326,7 @@ void MTrk::ImportXmiEVNT(const Chunk & evnt)
 	}
 
 	// read delta
-	while(0 < *ptr)
+	while(0 < static_cast<s8>(*ptr))
 	{
 	    delta += *ptr;
 	    ++ptr;
